fix: correct release.sh script

+++ release.sh
@@ -1,9 +1,18 @@
 #!/bin/bash
 set -ev
+
+if [ $TRAVIS_PULL_REQUEST == "true" ]; then
+  echo "this is PR, exiting"
+  exit 0
+fi
+
 rm -fr ./node_modules
 git config --global user.name 'standard-version'
 git config --global user.email 'standard-version@travis'
 git remote "https://${GITHUB_TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git"
+
+git config credential.helper "store --file=.git/credentials"
+echo "https://${GITHUB_TOKEN}:@github.com" > .git/credentials
 
 echo $TRAVIS_BRANCH
 echo $TRAVIS_COMMIT_RANGE
